back-end/requirements.txt
fastapi[standard]
uvicorn
pymongo
python-dotenv
redis
passlib[bcrypt]
bcrypt<5
python-jose
docker
itsdangerous